Suppport for nested affiliations.
Note, in each instance of a class, any attribute that is set needs to be inistaiated to it's default value
or it seems that another instance of the class might access the globally set value.

For example we are setting affiliations and author lists to zero on each instansiation of author and article respectivly.

Also self.orcid.set("xlink:href", contributor.orcid) returns an error

in aff, determine why some elements take an enclosing addr-line, and others don't

in aff, if email is associated with aff, how do we deal with two atuhors from the same place,
but with different emails?

Think about moving the function that adds the doctype out of the funciton that
does pretty printing.