Pinning script
This script will be useful for the bot to specify a fixed version of elife-poa-xml-generation. It will be called by elife-bot-formula specifying the path of a file contained in the elife-bot repository (similar to Lax's equivalent for bot-lax-adaptor, https://github.com/elifesciences/lax/blob/develop/bot-lax-adaptor.sha1).

This way, the bot carries together a fixed version of
elife-poa-xml-generation, that gets tested in ci and end2end along with
it.